fix(adcs): display HRESULT as unsigned hex in DCOM error messages
Cast act.HResult (int32) to uint32 before formatting with %08x
to prevent negative hex output like 0x-7ffbfeac instead of 0x80040154.

+++ Payload_Type/fawkes/fawkes/agent_code/pkg/commands/adcs_request.go
@@ -400,7 +400,7 @@ func adcsSubmitCSR(ctx context.Context, server, caName, template, altName string
 		return nil, fmt.Errorf("RemoteActivation: %v", err)
 	}
 	if act.HResult != 0 {
-		return nil, fmt.Errorf("RemoteActivation HRESULT: 0x%08x", act.HResult)
+		return nil, fmt.Errorf("RemoteActivation HRESULT: 0x%08x", uint32(act.HResult))
 	}
 
 	// Step 4: Dial the OXID endpoint for the activated object
@@ -505,7 +505,7 @@ func adcsQueryEditFlags(ctx context.Context, server, caName string, cred sspcred
 		return 0, fmt.Errorf("RemoteActivation: %v", err)
 	}
 	if act.HResult != 0 {
-		return 0, fmt.Errorf("RemoteActivation HRESULT: 0x%08x", act.HResult)
+		return 0, fmt.Errorf("RemoteActivation HRESULT: 0x%08x", uint32(act.HResult))
 	}
 
 	// Dial OXID endpoint
